+++ base/src/main/kotlin/com/commonsense/android/kotlin/base/algorithms/RunningAverage.kt
@@ -2,11 +2,11 @@ package com.commonsense.android.kotlin.base.algorithms
 
 
 /**
- * Created by Kasper Tvede on 28-04-2018.
- * Purpose:
+ * Created by Kasper Tvede
+ * The base class for handling calculating the average, with the ability to add more values to it
+ * /akk the running average.
  *
  */
-
 abstract class RunningAverageAbstract<T : Number> {
 
     private var numberCount: Long = 0