+++ system/src/main/kotlin/com/commonsense/android/kotlin/system/base/interfaceBases/BaseActivityLifecycleCallbacks.kt
@@ -1,0 +1,39 @@
+package com.commonsense.android.kotlin.system.base.interfaceBases
+
+import android.app.*
+import android.app.Application.*
+import android.os.*
+
+/**
+ *
+ */
+
+interface BaseActivityLifecycleCallbacks : ActivityLifecycleCallbacks {
+    override fun onActivityPaused(activity: Activity?) {
+
+    }
+
+    override fun onActivityResumed(activity: Activity?) {
+
+    }
+
+    override fun onActivityStarted(activity: Activity?) {
+
+    }
+
+    override fun onActivityDestroyed(activity: Activity?) {
+
+    }
+
+    override fun onActivitySaveInstanceState(activity: Activity?, outState: Bundle?) {
+
+    }
+
+    override fun onActivityStopped(activity: Activity?) {
+
+    }
+
+    override fun onActivityCreated(activity: Activity?, savedInstanceState: Bundle?) {
+
+    }
+}